Added User Interactor
Added Block User Interactor, adds blocked user to user's block list and unfollows both users. Changed GetRelevantConversations to now filter through any conversations authored by users on an individuals block list so that they are not presented conversations authored by blocked users.

Also made some changes to Conversation and User to implement block lists (User) and added a getAuthor method to more easily fetch the author of a Conversation.

Finally I changed the delete user interactor to change "Authorship" of a conversation to the next user in the user list if a user deletes their account (who was previously the author to a conversation).